Fix issues caused by git merge related to NIRReflectance
