+++ satpy/tests/test_composites.py
@@ -453,244 +453,6 @@ class TestInlineComposites(unittest.TestCase):
                          ['IR_108', 'IR_087'])
 
 
-class TestNIRReflectance(unittest.TestCase):
-    """Test NIR reflectance compositor."""
-
-    def setUp(self):
-        """Set up the test case for the NIRReflectance compositor."""
-        self.get_lonlats = mock.MagicMock()
-        self.lons, self.lats = 1, 2
-        self.get_lonlats.return_value = (self.lons, self.lats)
-        area = mock.MagicMock(get_lonlats=self.get_lonlats)
-
-        self.start_time = 1
-        self.metadata = {'platform_name': 'Meteosat-11',
-                         'sensor': 'seviri',
-                         'name': 'IR_039',
-                         'area': area,
-                         'start_time': self.start_time}
-
-        nir_arr = np.random.random((2, 2))
-        self.nir = xr.DataArray(da.from_array(nir_arr), dims=['y', 'x'])
-        self.nir.attrs.update(self.metadata)
-
-        ir_arr = 100 * np.random.random((2, 2))
-        self.ir_ = xr.DataArray(da.from_array(ir_arr), dims=['y', 'x'])
-        self.ir_.attrs['area'] = area
-
-        self.sunz_arr = 100 * np.random.random((2, 2))
-        self.sunz = xr.DataArray(da.from_array(self.sunz_arr), dims=['y', 'x'])
-        self.sunz.attrs['standard_name'] = 'solar_zenith_angle'
-        self.sunz.attrs['area'] = area
-        self.da_sunz = da.from_array(self.sunz_arr)
-
-        refl_arr = np.random.random((2, 2))
-        self.refl = da.from_array(refl_arr)
-        self.refl_with_co2 = da.from_array(np.random.random((2, 2)))
-        self.refl_from_tbs = mock.MagicMock()
-        self.refl_from_tbs.side_effect = self.fake_refl_from_tbs
-
-    def fake_refl_from_tbs(self, sun_zenith, da_nir, da_tb11, tb_ir_co2=None):
-        """Fake refl_from_tbs."""
-        del sun_zenith, da_nir, da_tb11
-        if tb_ir_co2 is not None:
-            return self.refl_with_co2
-        else:
-            return self.refl
-
-    @mock.patch('satpy.composites.sun_zenith_angle')
-    @mock.patch('satpy.composites.NIRReflectance.apply_modifier_info')
-    @mock.patch('satpy.composites.Calculator')
-    def test_provide_sunz_no_co2(self, calculator, apply_modifier_info, sza):
-        """Test NIR reflectance compositor provided only sunz."""
-        calculator.return_value = mock.MagicMock(
-            reflectance_from_tbs=self.refl_from_tbs)
-        sza.return_value = self.da_sunz
-        from satpy.composites import NIRReflectance
-
-        comp = NIRReflectance(name='test')
-        info = {'modifiers': None}
-        res = comp([self.nir, self.ir_], optional_datasets=[self.sunz], **info)
-
-        assert self.metadata.items() <= res.attrs.items()
-        assert res.attrs['units'] == '%'
-        assert res.attrs['sun_zenith_threshold'] is not None
-        assert np.allclose(res.data, self.refl * 100).compute()
-
-    @mock.patch('satpy.composites.sun_zenith_angle')
-    @mock.patch('satpy.composites.NIRReflectance.apply_modifier_info')
-    @mock.patch('satpy.composites.Calculator')
-    def test_no_sunz_no_co2(self, calculator, apply_modifier_info, sza):
-        """Test NIR reflectance compositor with minimal parameters."""
-        calculator.return_value = mock.MagicMock(
-            reflectance_from_tbs=self.refl_from_tbs)
-        sza.return_value = self.da_sunz
-        from satpy.composites import NIRReflectance
-
-        comp = NIRReflectance(name='test')
-        info = {'modifiers': None}
-        res = comp([self.nir, self.ir_], optional_datasets=[], **info)
-
-        self.get_lonlats.assert_called()
-        sza.assert_called_with(self.start_time, self.lons, self.lats)
-        self.refl_from_tbs.assert_called_with(self.da_sunz, self.nir.data, self.ir_.data, tb_ir_co2=None)
-        assert np.allclose(res.data, self.refl * 100).compute()
-
-    @mock.patch('satpy.composites.sun_zenith_angle')
-    @mock.patch('satpy.composites.NIRReflectance.apply_modifier_info')
-    @mock.patch('satpy.composites.Calculator')
-    def test_no_sunz_with_co2(self, calculator, apply_modifier_info, sza):
-        """Test NIR reflectance compositor provided extra co2 info."""
-        calculator.return_value = mock.MagicMock(
-            reflectance_from_tbs=self.refl_from_tbs)
-        from satpy.composites import NIRReflectance
-        sza.return_value = self.da_sunz
-
-        comp = NIRReflectance(name='test')
-        info = {'modifiers': None}
-        co2_arr = np.random.random((2, 2))
-        co2 = xr.DataArray(da.from_array(co2_arr), dims=['y', 'x'])
-        co2.attrs['wavelength'] = [12.0, 13.0, 14.0]
-        co2.attrs['units'] = 'K'
-        res = comp([self.nir, self.ir_], optional_datasets=[co2], **info)
-
-        self.refl_from_tbs.assert_called_with(self.da_sunz, self.nir.data, self.ir_.data, tb_ir_co2=co2.data)
-        assert np.allclose(res.data, self.refl_with_co2 * 100).compute()
-
-    @mock.patch('satpy.composites.sun_zenith_angle')
-    @mock.patch('satpy.composites.NIRReflectance.apply_modifier_info')
-    @mock.patch('satpy.composites.Calculator')
-    def test_provide_sunz_and_threshold(self, calculator, apply_modifier_info, sza):
-        """Test NIR reflectance compositor provided sunz and a sunz threshold."""
-        calculator.return_value = mock.MagicMock(
-            reflectance_from_tbs=self.refl_from_tbs)
-        from satpy.composites import NIRReflectance
-        sza.return_value = self.da_sunz
-
-        comp = NIRReflectance(name='test', sunz_threshold=84.0)
-        info = {'modifiers': None}
-        res = comp([self.nir, self.ir_], optional_datasets=[self.sunz], **info)
-
-        self.assertEqual(res.attrs['sun_zenith_threshold'], 84.0)
-        calculator.assert_called_with('Meteosat-11', 'seviri', 'IR_039',
-                                      sunz_threshold=84.0, masking_limit=NIRReflectance.MASKING_LIMIT)
-
-    @mock.patch('satpy.composites.sun_zenith_angle')
-    @mock.patch('satpy.composites.NIRReflectance.apply_modifier_info')
-    @mock.patch('satpy.composites.Calculator')
-    def test_sunz_threshold_default_value_is_not_none(self, calculator, apply_modifier_info, sza):
-        """Check that sun_zenith_threshold is not None."""
-        from satpy.composites import NIRReflectance
-
-        comp = NIRReflectance(name='test')
-        info = {'modifiers': None}
-        calculator.return_value = mock.MagicMock(
-            reflectance_from_tbs=self.refl_from_tbs)
-        comp([self.nir, self.ir_], optional_datasets=[self.sunz], **info)
-
-        assert comp.sun_zenith_threshold is not None
-
-    @mock.patch('satpy.composites.sun_zenith_angle')
-    @mock.patch('satpy.composites.NIRReflectance.apply_modifier_info')
-    @mock.patch('satpy.composites.Calculator')
-    def test_provide_masking_limit(self, calculator, apply_modifier_info, sza):
-        """Test NIR reflectance compositor provided sunz and a sunz threshold."""
-        calculator.return_value = mock.MagicMock(
-            reflectance_from_tbs=self.refl_from_tbs)
-        from satpy.composites import NIRReflectance
-        sza.return_value = self.da_sunz
-
-        comp = NIRReflectance(name='test', masking_limit=None)
-        info = {'modifiers': None}
-        res = comp([self.nir, self.ir_], optional_datasets=[self.sunz], **info)
-
-        self.assertIsNone(res.attrs['sun_zenith_masking_limit'])
-        calculator.assert_called_with('Meteosat-11', 'seviri', 'IR_039',
-                                      sunz_threshold=NIRReflectance.TERMINATOR_LIMIT, masking_limit=None)
-
-    @mock.patch('satpy.composites.sun_zenith_angle')
-    @mock.patch('satpy.composites.NIRReflectance.apply_modifier_info')
-    @mock.patch('satpy.composites.Calculator')
-    def test_masking_limit_default_value_is_not_none(self, calculator, apply_modifier_info, sza):
-        """Check that sun_zenith_threshold is not None."""
-        from satpy.composites import NIRReflectance
-
-        comp = NIRReflectance(name='test')
-        info = {'modifiers': None}
-        calculator.return_value = mock.MagicMock(
-            reflectance_from_tbs=self.refl_from_tbs)
-        comp([self.nir, self.ir_], optional_datasets=[self.sunz], **info)
-
-        assert comp.masking_limit is not None
-
-
-class TestNIREmissivePartFromReflectance(unittest.TestCase):
-    """Test the NIR Emissive part from reflectance compositor."""
-
-    @mock.patch('satpy.composites.sun_zenith_angle')
-    @mock.patch('satpy.composites.NIREmissivePartFromReflectance.apply_modifier_info')
-    @mock.patch('satpy.composites.Calculator')
-    def test_compositor(self, calculator, apply_modifier_info, sza):
-        """Test the NIR emissive part from reflectance compositor."""
-        import numpy as np
-        import xarray as xr
-        import dask.array as da
-        from satpy.composites import NIRReflectance
-
-        refl_arr = np.random.random((2, 2))
-        refl = da.from_array(refl_arr)
-
-        refl_from_tbs = mock.MagicMock()
-        refl_from_tbs.return_value = refl
-        calculator.return_value = mock.MagicMock(reflectance_from_tbs=refl_from_tbs)
-
-        emissive_arr = np.random.random((2, 2))
-        emissive = da.from_array(emissive_arr)
-        emissive_part = mock.MagicMock()
-        emissive_part.return_value = emissive
-        calculator.return_value = mock.MagicMock(emissive_part_3x=emissive_part)
-
-        from satpy.composites import NIREmissivePartFromReflectance
-
-        comp = NIREmissivePartFromReflectance(name='test', sunz_threshold=86.0)
-        info = {'modifiers': None}
-
-        platform = 'NOAA-20'
-        sensor = 'viirs'
-        chan_name = 'M12'
-
-        get_lonlats = mock.MagicMock()
-        lons, lats = 1, 2
-        get_lonlats.return_value = (lons, lats)
-        area = mock.MagicMock(get_lonlats=get_lonlats)
-
-        nir_arr = np.random.random((2, 2))
-        nir = xr.DataArray(da.from_array(nir_arr), dims=['y', 'x'])
-        nir.attrs['platform_name'] = platform
-        nir.attrs['sensor'] = sensor
-        nir.attrs['name'] = chan_name
-        nir.attrs['area'] = area
-        ir_arr = np.random.random((2, 2))
-        ir_ = xr.DataArray(da.from_array(ir_arr), dims=['y', 'x'])
-        ir_.attrs['area'] = area
-
-        sunz_arr = 100 * np.random.random((2, 2))
-        sunz = xr.DataArray(da.from_array(sunz_arr), dims=['y', 'x'])
-        sunz.attrs['standard_name'] = 'solar_zenith_angle'
-        sunz.attrs['area'] = area
-        sunz2 = da.from_array(sunz_arr)
-        sza.return_value = sunz2
-
-        res = comp([nir, ir_], optional_datasets=[sunz], **info)
-        self.assertEqual(res.attrs['sun_zenith_threshold'], 86.0)
-        self.assertEqual(res.attrs['units'], 'K')
-        self.assertEqual(res.attrs['platform_name'], platform)
-        self.assertEqual(res.attrs['sensor'], sensor)
-        self.assertEqual(res.attrs['name'], chan_name)
-        calculator.assert_called_with('NOAA-20', 'viirs', 'M12', sunz_threshold=86.0,
-                                      masking_limit=NIRReflectance.MASKING_LIMIT)
-
-
 class TestColormapCompositor(unittest.TestCase):
     """Test the ColormapCompositor."""
