+++ satpy/tests/test_modifiers.py
@@ -222,7 +222,8 @@ class TestNIRReflectance(unittest.TestCase):
         res = comp([self.nir, self.ir_], optional_datasets=[self.sunz], **info)
 
         self.assertEqual(res.attrs['sun_zenith_threshold'], 84.0)
-        calculator.assert_called_with('Meteosat-11', 'seviri', 'IR_039', sunz_threshold=84.0)
+        calculator.assert_called_with('Meteosat-11', 'seviri', 'IR_039',
+                                      sunz_threshold=84.0, masking_limit=NIRReflectance.MASKING_LIMIT)
 
     @mock.patch('satpy.modifiers.spectral.sun_zenith_angle')
     @mock.patch('satpy.modifiers.NIRReflectance.apply_modifier_info')
@@ -231,13 +232,46 @@ class TestNIRReflectance(unittest.TestCase):
         """Check that sun_zenith_threshold is not None."""
         from satpy.modifiers.spectral import NIRReflectance
 
-        comp = NIRReflectance(name='test', sunz_threshold=None)
+        comp = NIRReflectance(name='test')
         info = {'modifiers': None}
         calculator.return_value = mock.MagicMock(
             reflectance_from_tbs=self.refl_from_tbs)
         comp([self.nir, self.ir_], optional_datasets=[self.sunz], **info)
 
         assert comp.sun_zenith_threshold is not None
+
+    @mock.patch('satpy.modifiers.spectral.sun_zenith_angle')
+    @mock.patch('satpy.modifiers.NIRReflectance.apply_modifier_info')
+    @mock.patch('satpy.modifiers.spectral.Calculator')
+    def test_provide_masking_limit(self, calculator, apply_modifier_info, sza):
+        """Test NIR reflectance compositor provided sunz and a sunz threshold."""
+        calculator.return_value = mock.MagicMock(
+            reflectance_from_tbs=self.refl_from_tbs)
+        from satpy.modifiers.spectral import NIRReflectance
+        sza.return_value = self.da_sunz
+
+        comp = NIRReflectance(name='test', masking_limit=None)
+        info = {'modifiers': None}
+        res = comp([self.nir, self.ir_], optional_datasets=[self.sunz], **info)
+
+        self.assertIsNone(res.attrs['sun_zenith_masking_limit'])
+        calculator.assert_called_with('Meteosat-11', 'seviri', 'IR_039',
+                                      sunz_threshold=NIRReflectance.TERMINATOR_LIMIT, masking_limit=None)
+
+    @mock.patch('satpy.modifiers.spectral.sun_zenith_angle')
+    @mock.patch('satpy.modifiers.NIRReflectance.apply_modifier_info')
+    @mock.patch('satpy.modifiers.spectral.Calculator')
+    def test_masking_limit_default_value_is_not_none(self, calculator, apply_modifier_info, sza):
+        """Check that sun_zenith_threshold is not None."""
+        from satpy.modifiers.spectral import NIRReflectance
+
+        comp = NIRReflectance(name='test')
+        info = {'modifiers': None}
+        calculator.return_value = mock.MagicMock(
+            reflectance_from_tbs=self.refl_from_tbs)
+        comp([self.nir, self.ir_], optional_datasets=[self.sunz], **info)
+
+        assert comp.masking_limit is not None
 
 
 class TestNIREmissivePartFromReflectance(unittest.TestCase):
@@ -251,6 +285,7 @@ class TestNIREmissivePartFromReflectance(unittest.TestCase):
         import numpy as np
         import xarray as xr
         import dask.array as da
+        from satpy.modifiers.spectral import NIRReflectance
 
         refl_arr = np.random.random((2, 2))
         refl = da.from_array(refl_arr)
@@ -302,7 +337,8 @@ class TestNIREmissivePartFromReflectance(unittest.TestCase):
         self.assertEqual(res.attrs['platform_name'], platform)
         self.assertEqual(res.attrs['sensor'], sensor)
         self.assertEqual(res.attrs['name'], chan_name)
-        calculator.assert_called_with('NOAA-20', 'viirs', 'M12', sunz_threshold=86.0)
+        calculator.assert_called_with('NOAA-20', 'viirs', 'M12', sunz_threshold=86.0,
+                                      masking_limit=NIRReflectance.MASKING_LIMIT)
 
 
 class TestPSPAtmosphericalCorrection(unittest.TestCase):
